Convert seed data from Ruby to yml
Note: In order to do this, I have to more some Ruby code that used to be in the
seed files somewhere else. Because of this, some of the Problem and Solution
data is replaced with strings that need to get converted to object lookups
somewhere else, like in a before save filter.